# vtysh: Speed up output of configuration across daemons